fix(ci): move isThisSemester out of EnrolledNowSection.tsx
react-refresh/only-export-components fails a component file that also
exports a plain function — CI's UI/UX gate caught it. Moved to the
existing SubjectsPanel/utils.ts (already the home for this file's
other pure helpers) and merged its tests into utils.test.ts instead of
leaving an orphaned test file behind.
