+++ src/components/SubjectsPanel/utils.ts
@@ -89,6 +89,28 @@ export function cleanGroupName(name: string): string {
   return cleaned.charAt(0).toUpperCase() + cleaned.slice(1);
 }
 
+// Winter semester: Sep 1 – Jan 31. Summer semester: Feb 1 – Aug 31.
+// fulfillmentDate is DD.MM.YYYY in Czech (dot separator) but MM/DD/YYYY in
+// English (slash separator, US month-first order) — verified against real
+// IS Mendelu output for the same record: CZ "14.01.2026" == EN "01/14/2026".
+export function isThisSemester(fulfillmentDate?: string): boolean {
+  if (!fulfillmentDate) return false;
+  const isSlash = fulfillmentDate.includes('/');
+  const parts = fulfillmentDate.split(/[./]/);
+  if (parts.length < 3) return false;
+  const [a, b, year] = parts;
+  const [day, month] = isSlash ? [b, a] : [a, b];
+  const date = new Date(Number(year), Number(month) - 1, Number(day));
+  const now = new Date();
+  const isSummer = now.getMonth() >= 1 && now.getMonth() <= 7;
+  const semStart = isSummer
+    ? new Date(now.getFullYear(), 1, 1)
+    : now.getMonth() >= 8
+      ? new Date(now.getFullYear(), 8, 1)
+      : new Date(now.getFullYear() - 1, 8, 1);
+  return date >= semStart && date <= now;
+}
+
 export function shortenStatusText(text: string): string {
   if (!text) return '';
   const clean = text.trim();
